+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 yarn.lock
 reports
 .stryker-tmp
+package-lock.json
