+++ plugins/css-blank-pseudo/.gitignore
@@ -1,14 +1,10 @@
 node_modules
-/browser.*
-/browser-legacy.*
-/index.*
-/legacy.*
-/postcss.*
+dist
 package-lock.json
 yarn.lock
 *.log*
 *.result.css
-.*
+*.result.css.map
 !.editorconfig
 !.gitignore
 !.rollup.js
