+++ plugins/css-blank-pseudo/README-POSTCSS.md
@@ -1,7 +1,6 @@
 # CSS Blank Pseudo for PostCSS [<img src="http://jonathantneal.github.io/postcss-logo.svg" alt="" width="90" height="90" align="right">][CSS Blank Pseudo]
 
 [![NPM Version][npm-img]][npm-url]
-[![Build Status][cli-img]][cli-url]
 [![Support Chat][git-img]][git-url]
 
 [CSS Blank Pseudo] lets you style form elements when they are empty, following
@@ -39,7 +38,7 @@ npm install css-blank-pseudo --save-dev
 Use [CSS Blank Pseudo] to process your CSS:
 
 ```js
-const postcssBlankPseudo = require('css-blank-pseudo/postcss');
+const postcssBlankPseudo = require('css-blank-pseudo');
 
 postcssBlankPseudo.process(YOUR_CSS /*, processOptions, pluginOptions */);
 ```
@@ -48,7 +47,7 @@ Or use it as a [PostCSS] plugin:
 
 ```js
 const postcss = require('postcss');
-const postcssBlankPseudo = require('css-blank-pseudo/postcss');
+const postcssBlankPseudo = require('css-blank-pseudo');
 
 postcss([
   postcssBlankPseudo(/* pluginOptions */)
@@ -69,7 +68,7 @@ The `preserve` option defines whether the original selector should remain. By
 default, the original selector is preserved.
 
 ```js
-focusWithin({ preserve: false });
+cssBlankPseudo({ preserve: false });
 ```
 
 ```css
@@ -90,7 +89,7 @@ The `replaceWith` option defines the selector to replace `:blank`. By
 default, the replacement selector is `[blank]`.
 
 ```js
-focusWithin({ replaceWith: '.blank' });
+cssBlankPseudo({ replaceWith: '.blank' });
 ```
 
 ```css
@@ -109,12 +108,10 @@ input:blank {
 }
 ```
 
-[cli-img]: https://img.shields.io/travis/csstools/css-blank-pseudo/master.svg
-[cli-url]: https://travis-ci.org/csstools/css-blank-pseudo
 [git-img]: https://img.shields.io/badge/support-chat-blue.svg
 [git-url]: https://gitter.im/postcss/postcss
 [npm-img]: https://img.shields.io/npm/v/css-blank-pseudo.svg
 [npm-url]: https://www.npmjs.com/package/css-blank-pseudo
 
-[CSS Blank Pseudo]: https://github.com/csstools/css-blank-pseudo
+[CSS Blank Pseudo]: https://github.com/csstools/postcss-plugins/tree/main/plugins/css-blank-pseudo
 [Selectors Level 4]: https://drafts.csswg.org/selectors-4/#blank
