+++ plugins/css-blank-pseudo/src/browser-global.js
@@ -1,0 +1,3 @@
+/* global self */
+import * as cssBlankPseudo from './browser';
+self.cssBlankPseudo = cssBlankPseudo;
