--- plugins/css-blank-pseudo/src/postcss.js
@@ -1,29 +1,0 @@
-const selectorRegExp = /:blank([^\w-]|$)/gi;
-
-const creator = opts => {
-	const replaceWith = String(Object(opts).replaceWith || '[blank]');
-	const preserve = Boolean('preserve' in Object(opts) ? opts.preserve : true);
-
-	return {
-		postcssPlugin: 'css-blank-pseudo',
-		Once: root => {
-			root.walkRules(selectorRegExp, rule => {
-				const selector = rule.selector.replace(selectorRegExp, ($0, $1) => {
-					return `${replaceWith}${$1}`;
-				});
-
-				const clone = rule.clone({ selector });
-
-				if (preserve) {
-					rule.before(clone);
-				} else {
-					rule.replaceWith(clone);
-				}
-			});
-		},
-	};
-};
-
-creator.postcss = true
-
-export default creator
