+++ plugins/css-blank-pseudo/test/basic.css
@@ -25,3 +25,7 @@ test :matches(test :blank :blank test) test {
 :blankignore {
 	order: 3;
 }
+
+test:not(:blank) {
+	order: 4;
+}
