+++ plugins/css-blank-pseudo/test/basic.expect.css
@@ -6,8 +6,7 @@
 	order: 1;
 }
 
-[blank],
-[blank] test,
+[blank],[blank] test,
 test [blank],
 test test[blank],
 test [blank] test,
@@ -45,4 +44,12 @@ test :matches(test :blank :blank test) test {
 :ignoreblank,
 :blankignore {
 	order: 3;
+}
+
+test:not([blank]) {
+	order: 4;
+}
+
+test:not(:blank) {
+	order: 4;
 }
